[back] feat: 챔피언/팀워크/맞라인 통계 정렬 기준 개선 및 타입 안정화
- 챔피언, 팀워크, 맞라인 통계 정렬 기준 세분화
  - 전적 많은 순 → 승률 높은 순
  - 승률 낮은 순 → 전적 많은 순 등
- 인기 챔피언 통계 정렬 조건 적용
- ChampionStatDto 필드 long 타입으로 유지 (형변환 제거)

+++ back/roflparser/src/main/java/com/example/roflparser/dto/response/ChampionStatisticsResponse.java
@@ -16,9 +16,9 @@ public class ChampionStatisticsResponse {
     @Builder
     public static class ChampionStatDto {
         private String name;
-        private int matches;
-        private int wins;
-        private int losses;
+        private long matches;
+        private long wins;
+        private long losses;
         private double winRate;
     }
 
@@ -26,9 +26,9 @@ public class ChampionStatisticsResponse {
     @Builder
     public static class ChampionScoreDto {
         private String name;
-        private int matches;
-        private int wins;
-        private int losses;
+        private long matches;
+        private long wins;
+        private long losses;
         private double winRate;
         private double pickRate;
         private double score;
